site: what if we generated a github site of mdbook
mdbook blows away its target dir, so... we just make a new repo, bulk add, and force push everytime.  because we can, that works, and idgaf about the history of the generated assets.  not sure if git is capable of object diffing sensible to reduce upload; it should be, but theory and practice might not be congruent here.

Signed-off-by: Eric Myhre <hash@exultant.us>