site: empty outline.
I'd like the markdown source to be in a dir, and not named 'src', but apparently we *have* to have a path named 'src' (and no, the arg to `mdbook build` doesn't do what you'd think, exactly) -- at least as of v0.0.21, which is what I'm using as its the most recent release binary on github as of $TODAY (despite the fact that yes, it's several months behind master; I don't have the wherewithall to recurse my problems into from-source builds of this tool today).  So, some symlinks appear.  Whatever.

Signed-off-by: Eric Myhre <hash@exultant.us>