TOC for CLI tools.
Note that yes, these links are insane.  mdbook appears to do paths... Wrong.
Evidentally it ignores all leading ".." segments, and it also treats "." as
the root of the site rather than relative to the page you're on.  I dunno.
Whatever; it does what it does and we'll just have to roll with it.

Signed-off-by: Eric Myhre <hash@exultant.us>