Update source hashes to latest; build omnibus.
This is destined to be release v0.201.

Stellar is now included; and like in the last release,
we have builds which operate *either* via the build.sh demo,
*or* via `stellar emerge`, and they will produce the same results.

Heft and hitch binaries are dropped from this release.  Stellar is
expected to grow the features replacing both of them, since the plan
for commissioning and releasing seems to be drawing all of these
features together very closely for getting a smooth control flow when
we're working on a really big picture.

Signed-off-by: Eric Myhre <hash@exultant.us>