Update module.tl for the Operation.Inputs flip.
This change was introduced in stellar 3128202 / go-timless-api cf6aecf.

Taking a little artistic license with the whitespace, as well.  It's
JSON; this is allowed.  And while typically, I don't think it a great
use of time to get jazzy with a document that's essentially supposed to
be machine readable / machine writeable... This one in particular is
often a point of reference for newcomers to the system, so, it's worth
a little more art than usual.

Also, drop old module symlink.  That form of file layout is now
thoroughly gone, as of the workspace rework in stellar e437e8f.

Signed-off-by: Eric Myhre <hash@exultant.us>